requirements.txt
requests==2.26.0
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
PyJWT==2.1.0
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3
simplejwt==2.0.1
